chore: relocation package
