--- src/main/java/dev/smjeon/commerce/oauth/SocialUserProperty.java
@@ -1,7 +1,0 @@
-package dev.smjeon.commerce.oauth;
-
-public interface SocialUserProperty {
-    String getNickName();
-
-    String getProfileUrl();
-}
