+++ src/main/java/dev/smjeon/commerce/oauth/common/SocialProviders.java
@@ -1,4 +1,4 @@
-package dev.smjeon.commerce.oauth;
+package dev.smjeon.commerce.oauth.common;
 
 public enum SocialProviders {
     KAKAO, GITHUB
