+++ src/main/java/dev/smjeon/commerce/oauth/github/application/GithubLoginService.java
@@ -1,9 +1,10 @@
-package dev.smjeon.commerce.oauth.application;
+package dev.smjeon.commerce.oauth.github.application;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import dev.smjeon.commerce.oauth.SocialUserInfo;
+import dev.smjeon.commerce.oauth.common.application.SocialLoginService;
+import dev.smjeon.commerce.oauth.common.dto.SocialUserInfo;
 import dev.smjeon.commerce.oauth.github.GithubConfig;
 import dev.smjeon.commerce.oauth.github.dto.GithubToken;
 import org.springframework.http.HttpHeaders;
